+++ inst/query.sql
@@ -1,3 +1,4 @@
 -- !preview conn=DBI::dbConnect(RSQLite::SQLite())
--- tar_load(query_params)
+-- tar_load(cars)
+-- tar_load(flowers)
 select 1 as my_col