+++ NEWS.md
@@ -1,4 +1,12 @@
-# sqltargets 0.1.1
+# sqltargets 0.2.0
+
+* Included 'jinjar' as a dependency, allowing jinja-like SQL queries.
+* New option `sqltargets.template_engine` (either 'glue' or 'jinjar').
+* Breaking change: `query_params` argument in `tar_sql()` is now `params`.
+* Additional tests
+* `params` object name now recognized as an upstream dependency.
+
+# sqltargets 0.1.0
 
 * Added two additional options for the opening and closing delimiters.
 * Upstream target name is now the file basename.